fix: ensure user ID comparison is type-safe in order and checkout success methods

+++ vcom/app/Http/Controllers/OrderController.php
@@ -14,7 +14,7 @@ class OrderController extends Controller
 
     public function show(Order $order)
     {
-        if ($order->user_id !== auth()->id()) {
+        if ((int) $order->user_id !== (int) auth()->id()) {
             abort(403);
         }
